Adding an old project to tidbits. The project has a few things that are good.
The project was many years ago to move data files from a storage area
that was stored in one format and convert it over to another format.
The process was very lengthy and should have been put into a fork, but
at the time I was not familar with forking in perl, so it was never put
in. The project was eventually ported over to another group that had the
correct permissions on the server to develop without asking an admin to
copy and run and copy and run to fix issues encountered. The good old
days of testing in production...